python-web-main/link_bio/requirements.txt
cloudpickle==2.2.1
configcat-client==9.0.2
h2==4.1.0
openai==1.42.0
python-dotenv==1.0.1
pytz==2024.1
supabase==2.3.5

reflex==0.5.10
